+++ src/app/api/execucaoplanejada/[id]/route.ts
@@ -9,11 +9,14 @@ export async function DELETE(
   try {
     const id = parseInt((await  params).id, 10);
     const userId = await verifyUser(req);
+
     if (userId instanceof NextResponse) return userId; 
+    if (isNaN(id)) return NextResponse.json({ message: "ID inválido" }, { status: 400 });
 
     const requestData = await req.json();
 
     const treinoId = parseInt(requestData.treinoId)
+    if (isNaN(treinoId)) return NextResponse.json({ message: "inválido" }, { status: 400 });
 
     await prisma.execucaoPlano.delete({
       where: {
@@ -29,7 +32,7 @@ export async function DELETE(
   } catch (error) {
     console.log(error);
     return NextResponse.json(
-      { message: "Erro ao excluir a treino" },
+      { message: "Erro ao excluir a exec" },
       { status: 500 }
     );
   }
